requirement.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Werkzeug==2.2.2
Flask-WTF==1.0.1
WTForms==3.0.1
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.83.0
python-dateutil==2.8.2